move createProClient to utils
